requirements.txt
Pytorch-Lightning==1.9.0
decord
kornia
timm
open_clip_torch
av
omegaconf
transformers
einops
scikit-learn
taming-transformers-rom1504
pandas
triton
fairscale
psutil==5.9.5
annotated-types==0.5.0
plotly
imageio==2.14.1
imageio-ffmpeg==0.4.7
opencv-python==4.8.0.74
moviepy
Pillow
tqdm
gradio==3.37.0